Add example of generator-driven menu commands
This showcases how we can have a non-Export attribute that
drives codegen that in turn creates the actual exports, which
can be smarter and provide more concise syntax. In this case,
we allow menu commands to be just any method in an
arbitrary class, which we then just invoke from the newly
created export.

Method parameters are considered imports, except for the
cancellation token (if any). This is the most intuitive approach,
since the IMenuCommand.ExecuteAsync interface has
only that as the (optional) parameter.

We should add support for annotating a class too,
and emitting the relevant partial class annotations in that
case (Export and ExportMetadata). It might seem useless
in that case, but that allows us to have a single attribute
for two very different (yet logically identical) concepts.