Add support for [MenuCommand] at the class level
Since the attribute does not inherit from ExportAttribute (so we can
apply it to methods with arbitrary signatures), we instead generate a
partial class with the relevant export and metadata attributes.

A new analyzer and code fixer makes sure the requirements are
satisfied, namely that the class is declared partial, and that it is
annotated with [MenuCommand]. This last requirement is made
a warning instead of an error, since it might be conceivable that one
might use our interface in other contexts (i.e. in tests?).