Make exports generator more flexible, support class and records
Since constructors for MEF need to be annotated with [ImportingConstructor] and that cannot be applied to a record primary constructor (since there's no `ctor` target, unlike `property`), we need to also support a regular class with a regular constructor, in addition to partial types.

We also add the originating assembly for an export as metadata, so we can properly filter out incompatible types when getting exports.