Improve composition by moving type compatibility down
The composition manager/adapter itself does the export
type compability check instead, initially based on metadata
for faster filtering (and lazy-preserving behavior).

This behavior is required now that we have more pervasive
duck-typing, since otherwise MEF will happily resolve types
that are incompatible as long as their type identify (full name)
happens to be the same. This fails at run-time with an invalid
cast exception since the Lazy<T> constructed from the
requested T cannot be constructed with the located T
in the composition which happens to have the same
full name but cannot be directly cast to the target T.