Add support for INotifyPropertyChanged to IContext
Context activation/deactivation/changes will eventually
drive much of the UI, and we want that process to be as
painless as possible. Therefore, notifying property/context
changes is quite important.

This commit improves the existing implementation by
detecting whether the pushed object implements INPC
and subscribing automatically for re-publishing the event
at the IContext level.

Likewise, pushing/disposing contexts also raises changes,
but with coarse granularity (no way to tell what specifically
changed. This is intentional to keep the main scenario
clean and simple.

A binding will later leverage this for UI behaviors.

Closes #43.